+++ src/Mvc/Controller/Traits/Abstracts/AbstractModel.php
@@ -28,6 +28,8 @@ trait AbstractModel
     abstract public function getControllerName(): string;
     
     abstract public function loadModel(?string $modelName = null): \Phalcon\Mvc\ModelInterface;
+
+    abstract public function modelHasColumn(string $column, ?string $modelName = null): bool;
     
     abstract public function appendModelName(string $field, ?string $modelName = null): string;
 }